requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
camelcase==0.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
h11==0.14.0
idna==3.4
lxml @ file:///E:/task/scrapping/updated/lxml-4.9.0-cp311-cp311-win_amd64.whl
mysql-connector-python==8.0.31
numpy==1.23.5
outcome==1.2.0
protobuf==3.20.1
pycparser==2.21
PySocks==1.7.1
requests==2.28.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
